Log SSL protocol and cipher, host, request time
This unifies our logging format for proxy protocol and HTTP, and adds
the SSL protocol and cipher, as well as the request time. SSL parameters
will be helpful to help customers deprecate support for SSLv3 (and
ideally migrate to ALBs in the process), and the latter is something
we've often wanted to use in debugging.

This also adds test that actually check these fields are logged.

Note that the new log format isn't really compliant with the combined
log format anymore... but that probably doesn't matter.