Health check: set a X-Aptible-Health-Check header
This lets customers identify whether a request is coming from a health
check or not when it reaches `/healthcheck` on their app.

This is not entirely foolproof considering and end-user can still
request `/.aptible/alb-healthcheck` (for ALB Endpoints).

We should investigate the possibility of black-holing this at the ALB
level, and review whether the ALB actually sends a trustable header
flagging a health check request that we could use here. We could even
have the ALB use a randomized health check path that only works for one
given release. The latter would actually be reasonably easy to
implement, with the caveat that coordinating its release might be a
little complex.

We should probably work on this once we start supporting health checks
where the status code is meaningful (where customers may want decide to
perform infrastructure health checks that end-users should not be
allowed to trigger).