ALB: debounce health check requests
This extends a our LUA-based health check to debounce health checks by
caching the status of the last health check for 5 seconds, and waiting
on running health checks for up to 1 second (it's pretty common for
health checks to all come in rapid succession, so only caching wouldn't
be ideal here).

The goal here is to limit the number of health health checks the
underlying app containers see. For example, this can merge the HTTP and
HTTPS health checks for a single container into a single request.

As part of this, I also rewrote the dummy upstream we use in tests to a
Ruby threaded implementation as opposed to the sequential nc-based
implementation we had until now. The main benefit is that it's not racy:
if a request comes in before we handle the current one, we don't drop it
to the floor. This was necessary to exercise how debouncing works when
concurrent requests are coming in and we don't have a status cached.